From Stephan Huber, "thomas kirk schrieb: > I put this at the end of realizeImplementation; someone with better knowledge > of Carbon programming may see a more appropriate place for the call.
I moved your code into the ctor of the OSXCarbonWindowingSystemInterface
so it get called only once. Can you test it again, if it works on your side?

I also disabled multithreaded-rendering because it slowed down the
rendering on my machine by a factor of 3. Perhaps we can make it
optional to test it on other machines.

I had some problems implementing pbuffer-support for os x and stopped it
for now until I have more time to investigate the issues.
"


git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@6106 16af8721-9629-0410-8352-f15c8da7e697