From Charles Cole, "Attached are mods to the OpenFlight plug-in to help further implement the specification.  With these mods, blink sequences are now created for flashing light point nodes, either palletized (v.15.8 and later) or non-palletized (15.7 and earlier).  Thanks to Brede for his implementation of the palletized light point nodes.
There is still work to do on adding the capability to properly handle
light point system nodes, but this does add some capability that did not
previously exist.  So, I wanted to at least submit this and I will
hopefully provide the additional capability in the near future.

I've tested the code modifications with Visual Studio 2005.  I don't
have the means to test any other operating system, but I would suspect
that there shouldn't be any issue (famous last words).  I used the test
files that I uploaded to the users forum to test the changes.

In addition to the added capability, I changed the light point node
radius to the "actualPixelSize" value in the file.  Previously, the
radius was set to half the actual pixel size (see
LightPointRecords.cpp).  Not sure why this was the case.  But, it was
brought to my attention by a co-worker who created the OpenFlight files
and was testing them with different viewers.  If there's some history
for setting the radius to half the size, then this change can be
omitted."


git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@6759 16af8721-9629-0410-8352-f15c8da7e697