From Mike Wittman, "This change to genwrapper and osgIntrospection gives access to the declaring file for a given type via the new member function osgIntrospection::Type::getDeclaringFile.  This information is useful in order to know what header to include when auto-generating wrappers for a given type.
During the C# wrapper generator development I've been keeping the
declaring file configuration state up-to-date manually with changes
to OSG, and it's proven to require substantial effort.  So it would be
extremely valuable to get this change in before 2.0 to reduce maintenance
during the lifetime of the release.  It'll also be equally useful to
others looking to create wrapper generators using osgIntrospection.

This is a fairly simple change and was tested with a fresh rebuild of the
entire suite of osgWrapper libraries, so it should be relatively low risk
(fingers crossed)."




git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@6900 16af8721-9629-0410-8352-f15c8da7e697