From Mario Valle,
osgviewerWX: "To make the example compile using a wx build non UNICODE based.
Tested on Linux with wxGTK 2.8.4"

osgviewerFOX: "Added removeChore() call in the FOX_OSG_MDIView destructor to get rid of a Trace/BPT trap
error on exit on Linux. BTW this is suggested also in the FOX documentation."



git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@7283 16af8721-9629-0410-8352-f15c8da7e697