From Jose Delport: "have been using the new VIEW_DEPENDENT_WITH_ORTHOGRAPHIC_OVERLAY mode of the OverlayNode.
I change the overlay subgraph dynamically and when I remove all the
subgraph nodes that is inside the current main camera FOV (others
outside still exist), the overlay texture does not update because of the
early return in the traversal. I then get a kind of ghost texture moving
around the terrain.

The attached file fixed the problem for me, but I'm not sure if it is
the best way to address the problem."


git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@7457 16af8721-9629-0410-8352-f15c8da7e697