From Serge Lages, "Hi Robert,
You were right about the CMAKE_MODULE_LINKER_FLAGS option for CMake, so here is a modification allowing to not generate the manifest files for the plugins making them a lot more easy to redistribute. I have also made the same modification to the wrappers as they are also put into the osgPlugin folder when generated.
"


git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@7629 16af8721-9629-0410-8352-f15c8da7e697