From Wojciech Lewandowski, "By complete coincidence I just have read an old OSG submission mail related to bugfixes in osg::Polytope.setToUnitFrustum and setToBoundingBox  It was sent at beginning of december. I read it when purging my Thrash emails and found it there this because it was wrongly classified as SPAM.
What stroke me in this email was the fact that there was once an error in
Polytope class. Since I adopted CustomPolytope (osgSim OverlayNode.cpp) for
my minimal shadow area computations I checked my code for this error. And I
found it in CustomPolytope::setToUnitFrustum method.
CustomPolytope::setToBoundingBox seemed OK.

So I went back to the origin and fixed this error in OverlayNode.cpp as
well. I have not tested it in OverlayNode though (I don't know how) so
please look at this carefully. But it seems to work fine with my shadow
calculations."


git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@7718 16af8721-9629-0410-8352-f15c8da7e697