Refactored DatabasePager and related classes to introduce support for multi-threaded paging, where the Pager manages threads of reading local and http files via seperate threads.  This makes it possible to smoothly browse large databases where parts of the data are locally cached while others are on a remote server.  Previously with this type of dataset the pager would stall all paging while http requests were being served, even when parts of the models are still loadable virtue of being in the local cache.
Also as part of the refactoring the DatabaseRequest are now stored in the
ProxyNode/PagedLOD nodes to facilitate quite updating in the cull traversal,
with the new code avoiding mutex locks and searches.  Previous on big 
databases the overhead involved in make database requests could accumulate
to a point where it'd cause the cull traversal to break frame.  The overhead
now is negligable.

Finally OSG_FILE_CACHE support has been moved from the curl plugin into
the DatabasePager.  Eventually this functionality will be moved out into
osgDB for more general usage.   



git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@8325 16af8721-9629-0410-8352-f15c8da7e697